Added index checking to migration and autoupdate unit tests
Unit tests added by @raymondfeng's request on strongloop#64

Index ordering (ASC, DESC) works, but it currently isn't checked
to determine if an index needs to be recreated, that should be
relatively simple to add.

Index checks also need to be incorporated into isActual().

addIndexes() should maybe be changed to getAddModifyDropIndexes()
to conform with the getAddModifyColumns() and getDropColumns()
to allow usage to be more uniform. This would be helpful in isActual().
The SQL queries could just be returned rather than executed at the
end, so this should be an easy change if needed.

Signed-off-by: Rob Eroh <rob@eroh.me>